+++ src/libstdcrypto/sha1_update.c
@@ -1,10 +1,11 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/sha1.h>
 #include "sha1-internal.h"
 
-void sha1_update (SHA1Schedule *ctx, char const *buf, unsigned int len)
+void sha1_update (SHA1Schedule *ctx, char const *buf, size_t len)
 {
-  register unsigned int i = 0 ;
+  register size_t i = 0 ;
   for (; i < len ; i++) sha1_feed(ctx, (unsigned char)buf[i]) ;
 }
