+++ src/libstdcrypto/sha256_update.c
@@ -1,10 +1,11 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/sha256.h>
 #include "sha256-internal.h"
 
-void sha256_update (SHA256Schedule *ctx, char const *buf, unsigned int len)
+void sha256_update (SHA256Schedule *ctx, char const *buf, size_t len)
 {
-  register unsigned int i = 0 ;
+  register size_t i = 0 ;
   for (; i < len ; i++) sha256_feed(ctx, (unsigned char)buf[i]) ;
 }
